PEP-8 cleanup and reorganization.
