+++ fs9721/client.py
@@ -1,17 +1,20 @@
 import serial
 
+from .value import DmmValue
+from .exceptions import DmmNoData, DmmInvalidSyncValue, DmmReadFailure
 
-class Dmm:
+
+class Client(object):
     """
-    Takes readings off the serial port from a class of multimeters that includes
-    the TekPower TP4000ZC (the meter I own) and supposedly is the same as the the
-    'VC820' mode in QtDMM.
+    Takes readings off the serial port from a class of multimeters that
+    includes the TekPower TP4000ZC (the meter I own) and supposedly is the
+    same as the the 'VC820' mode in QtDMM.
 
     example code:
 
     # load the module
     import tp4000zc
-    
+
     # the port that we're going to use.  This can be a number or device name.
     # on linux or posix systems this will look like /dev/tty2 or /dev/ttyUSB0
     # on windows this will look something like COM3
@@ -22,7 +25,7 @@ class Dmm:
 
     # read a value
     val = dmm.read()
-    
+
     print val.text       # print the text representation of the value
                          # something like: -4.9 millivolts DC
     print val.numericVal # and the numeric value
@@ -33,8 +36,8 @@ class Dmm:
 
     Public Interface:
     __init__(port, retries=3, timeout=3.0):
-        Instantiating the class attempts to open the serial port specified, 
-        initialize it and read enough from the serial port to synchronize 
+        Instantiating the class attempts to open the serial port specified,
+        initialize it and read enough from the serial port to synchronize
         the module with the start/end of a full reading.
 
     read():
@@ -44,33 +47,35 @@ class Dmm:
     close():
         Finally you can close the serial port connection with close()
 
-    Exceptions will be raised if 
+    Exceptions will be raised if
        * PySerial raises an exception (SerialException or ValueError)
-       * this module can't get a full reading that passes initial data integrity
-         checks (subclasses of DmmException)
+       * this module can't get a full reading that passes initial data
+         integrity checks (subclasses of DmmException)
        * I made a coding error (whatever python might throw)
 
     If no exceptions are raised the DmmValue might still fail various sanity
-    checks or not have a numeric value.  Ie I believe that showing showing 
+    checks or not have a numeric value.  Ie I believe that showing showing
     multiple decimal points makes no sense but is valid per the protocol so
     no exception is raised but the saneValue flag will be set to False in the
     DmmValue.
 
     Meter Documentation:
 
-    Per the documentation page, the meter spits out readings which are bursts of 
-    14 bytes every .25 seconds.  The high nibble of each byte is the byte number 
-    (1-14) for synchronization and sanity checks, the low nibble holds the data.
+    Per the documentation page, the meter spits out readings which are bursts
+    of 14 bytes every .25 seconds.  The high nibble of each byte is the byte
+    number (1-14) for synchronization and sanity checks, the low nibble holds
+    the data.
 
-    Each data bit represents an individual field on the LCD display of the meter, 
-    from segments of the 7 segment digits to individual flags.  Bytes 1 and 10-14
-    are flags (with four bits reserved/unmapped on this meter) and bytes (2,3), 
-    (4,5), (5,6) and (7,8) representing the individual digits on the display.
+    Each data bit represents an individual field on the LCD display of the
+    meter, from segments of the 7 segment digits to individual flags.  Bytes 1
+    and 10-14 are flags (with four bits reserved/unmapped on this meter) and
+    bytes (2,3), (4,5), (5,6) and (7,8) representing the individual digits on
+    the display.
 
-    For the digits, if the high bit of the first nibble of a digit is set then the
-    negative sign (for the first digit) or the leading decimal point is turned on.
-    the remaining bits of the two nibbles represent the elements of the 7 segment
-    digit display as follows:
+    For the digits, if the high bit of the first nibble of a digit is set then
+    the negative sign (for the first digit) or the leading decimal point is
+    turned on. the remaining bits of the two nibbles represent the elements of
+    the 7 segment digit display as follows:
 
       pos 1       nibble 1:   S123
      p     p      nibble 2:   4567
@@ -90,15 +95,17 @@ class Dmm:
 
     bytesPerRead = 14
 
-    def __init__(self, port='/dev/ttyUSB0', retries = 3, timeout = 3.0):
+    def __init__(self, port='/dev/ttyUSB0', retries=3, timeout=3.0):
         self.ser = serial.Serial(
-            port = port,
-            baudrate = 2400,
-            parity = serial.PARITY_NONE,
-            stopbits = serial.STOPBITS_ONE,
-            bytesize = serial.EIGHTBITS,
-            timeout = timeout)
-        self.retries = retries # the number of times it's allowed to retry to get a valid 14 byte read
+            port=port,
+            baudrate=2400,
+            parity=serial.PARITY_NONE,
+            stopbits=serial.STOPBITS_ONE,
+            bytesize=serial.EIGHTBITS,
+            timeout=timeout)
+
+        # the number of times it's allowed to retry to get a valid 14 byte read
+        self.retries = retries
 
         self._synchronize()
 
@@ -125,27 +132,25 @@ class Dmm:
             else:
                 success = True
                 break
-            
+
             # if we're here we need to resync and retry
             self._synchronize()
 
         if not success:
             raise DmmReadFailure()
-        
 
         val = ''
-        for (d1,d2,ch) in self.digits:
+        for (d1, d2, ch) in self.digits:
             highBit, digit = self._readDigit(bytes[d1-1], bytes[d2-1])
             if highBit:
                 val = val + ch
             val = val + digit
 
         attribs = self._initAttribs()
-        for k,v in self.bits.items():
+        for k, v in self.bits.items():
             self._readAttribByte(bytes[k-1], v, attribs)
 
         return DmmValue(val, attribs, readAttempt, bytes)
-                            
 
     def _synchronize(self):
         v = self.ser.read(1)
@@ -164,26 +169,61 @@ class Dmm:
             # read without throwing an exception so for now
             # I'll say no.
 
-
     bits = {
-        1: [('flags', 'AC'), ('flags', 'DC'), ('flags', 'AUTO'), ('flags', 'RS232')],
-        10:[('scale', 'micro'), ('scale', 'nano'), ('scale', 'kilo'), ('measure', 'diode')],
-        11:[('scale', 'milli'), ('measure', '% (duty-cycle)'), ('scale', 'mega'),
-            ('flags', 'beep')],
-        12:[('measure', 'Farads'), ('measure', 'Ohms'), ('flags', 'REL delta'),
-            ('flags', 'Hold')],
-        13:[('measure', 'Amps'), ('measure', 'volts'), ('measure', 'Hertz'),
-            ('other', 'other_13_1')],
-        14:[('other', 'other_14_4'), ('measure', 'degrees Celcius'), ('other', 'other_14_2'),
-            ('other', 'other_14_1')]}
+        1: [
+            ('flags', 'AC'),
+            ('flags', 'DC'),
+            ('flags', 'AUTO'),
+            ('flags', 'RS232')
+        ],
+        10: [
+            ('scale', 'micro'),
+            ('scale', 'nano'),
+            ('scale', 'kilo'),
+            ('measure', 'diode')
+        ],
+        11: [
+            ('scale', 'milli'),
+            ('measure', '% (duty-cycle)'),
+            ('scale', 'mega'),
+            ('flags', 'beep')
+        ],
+        12: [
+            ('measure', 'Farads'),
+            ('measure', 'Ohms'),
+            ('flags', 'REL delta'),
+            ('flags', 'Hold')
+        ],
+        13: [
+            ('measure', 'Amps'),
+            ('measure', 'volts'),
+            ('measure', 'Hertz'),
+            ('other', 'other_13_1')
+        ],
+        14: [
+            ('other', 'other_14_4'),
+            ('measure', 'degrees Celcius'),
+            ('other', 'other_14_2'),
+            ('other', 'other_14_1')
+        ]
+    }
 
-    digits = [(2,3,'-'), (4,5,'.'), (6,7,'.'), (8,9,'.')]
-    digitTable = {(0,5):'1', (5,11):'2', (1,15):'3', (2,7):'4', (3,14):'5',
-                  (7,14):'6', (1,5):'7', (7,15):'8', (3,15):'9', (7,13):'0',
-                  (6,8):'L', (0,0):' '}
+    digits = [
+        (2, 3, '-'), (4, 5, '.'), (6, 7, '.'), (8, 9, '.')
+    ]
+    digitTable = {
+        (0, 5): '1', (5, 11): '2', (1, 15): '3', (2, 7): '4', (3, 14): '5',
+        (7, 14): '6', (1, 5): '7', (7, 15): '8', (3, 15): '9', (7, 13): '0',
+        (6, 8): 'L', (0, 0): ' '
+    }
 
     def _initAttribs(self):
-        return {'flags':[], 'scale':[], 'measure':[], 'other':[]}
+        return {
+            'flags': [],
+            'scale': [],
+            'measure': [],
+            'other': []
+        }
 
     def _readAttribByte(self, byte, bits, attribs):
         b = ord(byte) % 16
@@ -202,158 +242,7 @@ class Dmm:
         b1 = b1 % 8
         b2 = ord(byte2) % 16
         try:
-            digit = self.digitTable[(b1,b2)]
+            digit = self.digitTable[(b1, b2)]
         except:
             digit = 'X'
         return highBit, digit
-            
-
-class DmmValue:
-    """
-    This is a representation of a single read from the multimeter.
-
-    Attributes in rough order of usefulness:
-    
-    Sanity checks:
-       saneValue: True if no sanity checks failed.
-    
-    High level computed fields:
-       text: Nicely formatted text representation of the value.
-       numericVal: numeric value after SI prefixes applied or None if value is non-numeric.
-       measurement: what is being measured.
-       delta: True if the meter is in delta mode.
-       ACDC: 'AC', 'DC' or None.
-       readErrors:  Number of failed reads attempts before successfully getting a reading 
-           from the meter.
-
-    Other, possibly useful, computed fields:
-       val: cleaned up display value
-       scale: SI prefix for val
-
-    Unprocessed values:
-       rawVal: Numeric display
-       flags: Various flags modifying the measurement
-       scaleFlags: SI scaling factor flags
-       measurementFlags: Flags to specify what the meter is measuring
-       reservedFlags: Flags that are undefined
-       rawBytes:  the raw, 14 byte bitstream that produced this value.
-    
-    """
-    def __init__(self, val, attribs, readErrors, rawBytes):
-        self.saneValue = True
-        self.rawVal = self.val = val
-        self.flags = attribs['flags']
-        self.scaleFlags = attribs['scale']
-        self.measurementFlags = attribs['measure']
-        self.reservedFlags = attribs['other']
-        self.readErrors = readErrors
-        self.rawBytes = rawBytes
-        self.text = 'Invalid Value'
-
-        self.processFlags()
-        self.processScale()
-        self.processMeasurement()
-        self.processVal()
-
-        if self.saneValue:
-            self.createTextExpression()
-
-    def createTextExpression(self):
-        text =  self.deltaText
-        text += self.val
-        text += ' '
-        text += self.scale
-        text += self.measurement
-        text += self.ACDCText
-        self.text = text
-
-    def processFlags(self):
-        flags = self.flags
-        self.ACDC = None
-        self.ACDCText = ''
-        self.delta = False
-        self.deltaText = ''
-
-        if 'AC' in flags and 'DC' in flags:
-            self.saneValue = False
-        if 'AC' in flags:
-            self.ACDC = 'AC'
-        if 'DC' in flags:
-            self.ACDC = 'DC'
-        if self.ACDC is not None:
-            self.ACDCText = ' ' + self.ACDC
-        if 'REL delta' in flags:
-            self.delta = True
-            self.deltaText = 'delta '
-
-    scaleTable = {'nano': 0.000000001, 'micro': 0.000001, 'milli': 0.001, 
-                  'kilo': 1000.0, 'mega': 1000000.0}
-    def processScale(self):
-        s = self.scaleFlags
-        self.scale = ''
-        self.multiplier = 1
-
-        if len(s) == 0:
-            return
-        if len(s) > 1:
-            self.saneValue = False
-            return
-        self.scale = s[0]
-        self.multiplier = self.scaleTable[self.scale]
-
-    def processMeasurement(self):
-        m = self.measurementFlags
-        self.measurement = None
-        if len(m) != 1:
-            self.saneValue = False
-            return
-        self.measurement = m[0]
-
-    def processVal(self):
-        v = self.rawVal
-        self.numericVal = None
-        if 'X' in v:
-            self.saneValue = False
-            return
-        if v.count('.') > 1:
-            self.saneValue = False
-            return
-
-        n = None
-        try:
-            n = float(v)
-        except:
-            pass
-
-        if n is not None:
-            self.val = '%s'%n  # this should remove leading zeros, spaces etc.
-            self.numericVal = n * self.multiplier
-
-    def __repr__(self):
-        return "<DmmValue instance: %s>"%self.text
-
-
-class DmmException:
-    "Base exception class for Dmm."
-
-class DmmNoData(DmmException):
-    "Read from serial port timed out with no bytes read."
-
-class DmmInvalidSyncValue(DmmException):
-    "Got an invalid byte during syncronization."
-
-class DmmReadFailure(DmmException):
-    "Unable to get a successful read within the number of allowed retries."
-
-
-def main():
-    dmm = Dmm()
-
-    while True:
-        val = dmm.read()
-        print val.text
-        print val.numericVal
-
-# main hook
-if __name__ == "__main__":
-    main()
