+++ requirements.txt
@@ -1,2 +1,4 @@
 measurement>=1.6
-pyserial==2.7
+pyserial>=2.7
+argparse>=1.2.1
+six
